cleanup and use doubles for matrix ops

+++ include/es/ffp.h
@@ -1,8 +1,7 @@
 #pragma once
 
-#include "glm/ext/matrix_float4x4.hpp"
+#include "glm/ext/matrix_double4x4.hpp"
 #include <GLES/gl.h>
-#include <memory>
 #include <stack>
 #include <vector>
 
@@ -15,5 +14,5 @@ inline GLenum currentPrimitive = GL_NONE;
 inline std::vector<GLfloat> floatVertexBuffer;
 
 inline GLenum currentMatrixMode = GL_MODELVIEW;
-inline glm::mat4 currentMatrix;
-inline std::stack<glm::mat4> matrixStack;
+inline glm::dmat4 currentMatrix;
+inline std::stack<glm::dmat4> matrixStack;
